# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MYiTH Core API
  description: |
    채용공고와 국가직무능력표준(NCS)을 결합한 개인 맞춤형 취업 로드맵 서비스.

    인증: Google ID Token으로 로그인 후 발급받은 accessToken을
    Authorization: Bearer {token} 헤더에 실어 보낸다.

    공통 응답: 모든 성공 응답은 data로 감싼다. (GET /api/health 제외)
    공통 오류: { "error": { code, message, fieldErrors, requestId } }
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: 로컬
- url: https://api.myith.example
  description: 운영
security:
- bearerAuth: []
...
paths:
...
  /api/characters:
    get:
      tags:
      - Roadmap
      summary: 캐릭터 목록 조회
      description: |-
        화면 1-1(알 선택, 캐릭터 없을 때)과 1-2(홈, 캐릭터 있을 때) 모두에 사용한다.
        프론트는 응답의 species 목록으로 알 선택 화면에서 보유 종류를 제외한다.
        nextQuest가 null이면 모든 퀘스트 완료 상태다.
        사이드바 카드와 홈 화면 카드가 함께 쓴다.
        status=archived로 아카이브된 로드맵을 조회한다(경험 카드 열람용).
      operationId: getCharacters
      parameters:
      - name: status
        in: query
        description: active | archived | all (기본 active)
        required: false
        schema:
          type: string
          enum:
          - active
          - archived
          - all
        example: active
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              example:
                data:
                - characterId: chr_01J3ABC
                  roadmapId: rmp_01J3ABC
                  species: deokbaseu
                  nickname: 견습 서버 개발자
                  jobCode: server
                  jobName: 백엔드 개발자
                  tagline: "서버와 데이터베이스를 설계하고, 안정적으로 동작하는 API를 구현합니다."
                  roadmapStatus: ACTIVE
                  completionRate: 80
                  stage: 4
                  stageLabel: 완성
                  level: 4
                  nextQuest:
                    questId: qst_01J3ABC
                    title: REST API 서버를 구현한다
                  createdAt: 2026-07-20T02:00:00Z
                  updatedAt: 2026-07-24T03:00:00Z
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT